move main code
